requirements.txt
django==1.11
redis==2.10.5
djangorestframework==3.6.2
django-redis==4.8.0
Celery==4.0.2
